frontend-backend app interface
- created app object to interact with backend, only state held is domainregistry artifact and selected account
