+++ frontend/src/index.html
@@ -1,7 +1,9 @@
 <!DOCTYPE html>
 <html>
 <head>
+  <title>BlindAuction</title>
 </head>
 <body>
+
 </body>
 </html>
